✅ Move test helpers to dedicated file
